+++ Runtime/Core/UIEnums.cs
@@ -57,7 +57,20 @@
         StartValue,
         CustomValue,
         OffsetFromStart,
-        OffsetFromCurrent
+        OffsetFromCurrent,
+        Direction
+    }
+
+    public enum UIAnimationDirection
+    {
+        Left = 0,
+        Right = 1,
+        Top = 2,
+        Bottom = 3,
+        TopLeft = 4,
+        TopRight = 5,
+        BottomLeft = 6,
+        BottomRight = 7
     }
 
     public enum UIPresetApplyMode